🌱 :: 현재 사용자 인증정보 가져오기

+++ DTL_Backend/src/main/java/com/example/dmstodo/controller/dto/req/MemberSignUpDto.java
@@ -21,7 +21,7 @@ public class MemberSignUpDto {
     @NotNull
     @Min(value = 1)
     @Max(value = 120)
-    private Long userAge;
+    private int userAge;
 
     @NotBlank
     @Size(max = 10)
